Port module account_banking_fr_lcr to v9
