+++ account_banking_fr_lcr/__openerp__.py
@@ -1,41 +1,21 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    French Letter of Change module for Odoo
-#    Copyright (C) 2014-2015 Akretion (http://www.akretion.com)
-#    @author: Alexis de Lattre <alexis.delattre@akretion.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# -*- coding: utf-8 -*-
+# © 2014-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
 
 {
     'name': 'French Letter of Change',
     'summary': 'Create French LCR CFONB files',
-    'version': '8.0.0.1.0',
+    'version': '9.0.1.0.0',
     'license': 'AGPL-3',
     'author': "Akretion,Odoo Community Association (OCA)",
     'website': 'http://www.akretion.com',
-    'category': 'Banking addons',
-    'depends': ['account_direct_debit', 'document'],
+    'category': 'French localisation',
+    'depends': ['account_payment_order', 'document'],
     'external_dependencies': {
         'python': ['unidecode'],
         },
-    'data': [
-        'wizard/export_lcr_view.xml',
-        'data/payment_type_lcr.xml',
-        ],
-    'demo': ['lcr_demo.xml'],
-    'installable': False,
+    'data': ['data/account_payment_method.xml'],
+    'demo': ['demo/lcr_demo.xml'],
+    'installable': True,
 }
