+++ account_banking_fr_lcr/data/account_payment_method.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+<data noupdate="1">
+
+
+<record id="fr_lcr" model="account.payment.method">
+    <field name="name">Lettre de Change Relevé</field>
+    <field name="code">fr_lcr</field>
+    <field name="payment_type">inbound</field>
+</record>
+
+
+</data>
+</odoo>
