--- account_banking_fr_lcr/data/payment_type_lcr.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data noupdate="1">
-
-<record id="payment_mode_type_lcr" model="payment.mode.type">
-    <field name="name">Lettre de Change Relevé</field>
-    <field name="code">LCR</field>
-    <field name="suitable_bank_types"
-        eval="[(6,0,[ref('base_iban.bank_iban')])]" />
-    <field name="ir_model_id" ref="model_banking_export_lcr_wizard"/>
-    <field name="payment_order_type">debit</field>
-</record>
-
-</data>
-</openerp>
