+++ account_banking_fr_lcr/demo/lcr_demo.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+<data noupdate="1">
+
+
+<record id="payment_mode_lcr" model="account.payment.mode">
+    <field name="name">LCR</field>
+    <field name="company_id" ref="base.main_company"/>
+    <field name="bank_account_link">variable</field>
+    <field name="payment_method_id" ref="fr_lcr"/>
+    <field name="default_journal_ids" search="[('type', 'in', ('sale', 'sale_refund'))]"/>
+</record>
+
+
+</data>
+</odoo>
