--- account_banking_fr_lcr/lcr_demo.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<openerp>
-<data noupdate="1">
-
-<record id="payment_mode_lcr" model="payment.mode">
-    <field name="name">LCR La Banque Postale</field>
-    <field name="journal" ref="account.bank_journal"/>
-    <field name="bank_id" ref="account_banking_payment_export.main_company_iban"/>
-    <field name="company_id" ref="base.main_company"/>
-    <field name="type" ref="payment_mode_type_lcr"/>
-</record>
-
-</data>
-</openerp>
