+++ account_banking_fr_lcr/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import account_payment_order
