--- account_banking_fr_lcr/wizard/export_lcr_view.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-  Copyright (C) 2014-2015 Akretion (http://www.akretion.com)
-  @author: Alexis de Lattre <alexis.delattre@akretion.com>
-  The licence is in the file __openerp__.py
--->
-
-<openerp>
-<data>
-
-<record id="banking_export_lcr_wizard_form" model="ir.ui.view">
-    <field name="name">banking.export.lcr.wizard.form</field>
-    <field name="model">banking.export.lcr.wizard</field>
-    <field name="arch" type="xml">
-        <form string="LCR File Generation">
-            <field name="state" invisible="1"/>
-            <group states="finish">
-                <field name="total_amount" />
-                <field name="file" filename="filename" />
-                <field name="filename" invisible="True"/>
-            </group>
-            <footer>
-                <button string="Generate" name="create_lcr" type="object"
-                        class="oe_highlight" states="create"/>
-                <button string="Validate" name="save_lcr" type="object"
-                        class="oe_highlight" states="finish"/>
-                <button string="Cancel" special="cancel"
-                        class="oe_link"/>
-            </footer>
-        </form>
-    </field>
-</record>
-
-</data>
-</openerp>
